backend/infrastructure/requirements.txt
asgiref==3.8.1
Django==4.2.19
djangorestframework==3.15.2
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==24.2
setuptools==75.8.0
sqlparse==0.5.3
tzdata==2025.1
